Do not rebuild containers when GitHub Actions are reconfigured
GHA configuration changes are not always related to container build
process. More often the reason would be (a) adding new container/tag or
(b) tweaking the pipeline itself.

This was a bad heuristic and it needs to be removed. Significant build
instructions are present only in Makefile anyways, so no lost capabilities
here